fuckkk

+++ tests/01-HeyTriangle/genShape.h
@@ -5,5 +5,5 @@ struct Geometry;
 
 Geometry makeNGon(size_t sides, float radius);
 
-Geometry makeCheckerBoard(int dim, int size);
+Geometry makeCheckerBoard(int dim, float size);
